chore: delete dead _shared/http.ts and its suite (live invariants re-homed); dedupe getBaseUrl; fix stale path comments
